flujo pop up arreglado
