+++ src/components/FormRegister/ModalRegister.jsx
@@ -23,16 +23,29 @@ import FB from "../FormLogin/svg/FB";
 import Gmail from "../FormLogin/svg/Gmail";
 import IconSch from "./svg/IconSch";
 import IconUser from "./svg/IconUser";
-const style1 = {
+import ModalRegistroFamilia from "./ModalRegistroFamilia";
+const styleDsktop = {
   position: "absolute",
-  top: "50%",
-  left: "50%",
+  top: "16%",
+  right: "-5%",
   transform: "translate(-50%, -50%)",
 
   //   maxWidth: 400,
   bgcolor: "background.paper",
   boxShadow: " 0px 1px 5px rgba(0, 0, 0, 0.40)",
-  padding: '2vh',
+  padding: "2vh",
+  borderRadius: "1vh",
+};
+const styleMobile = {
+  position: "absolute",
+  top: "50%",
+  right: "50%",
+  transform: "translate(-50%, -50%)",
+
+  //   maxWidth: 400,
+  bgcolor: "background.paper",
+  boxShadow: " 0px 1px 5px rgba(0, 0, 0, 0.40)",
+  padding: "2vh",
   borderRadius: "1vh",
 };
 
@@ -47,45 +60,7 @@ export default function ModalRegistro({ open, setOpen }) {
   const { distrits } = useSelector((state) => state.schools);
   const dispatch = useDispatch();
 
-  const ToggleSeePass = () => {
-    setseePassword(!seePassword);
-  };
-  const handlerOpenLogin = () => {
-    setOpenLogin(true);
-    // setOpen(false)
-    setOpenRegistroColegio(false);
-    setOpenRegistroPadre(false);
-  };
-
-  const {
-    register,
-    handleSubmit,
-    watch,
-    formState: { errors },
-  } = useForm({
-    defaultValues: {
-      name: "",
-      mail: "",
-      lastname: "",
-      phone: "",
-      password: "",
-    },
-    mode: "onChange",
-  });
   const navigate = useNavigate();
-
-  const OnSubmit = (user) => {
-    const data = {
-      apellidos: user.lastname,
-      email: user.mail,
-      nombre: user.name,
-      password: user.password,
-      telefono: user.phone,
-    };
-    console.log(data);
-    dispatch(registerUser(data));
-    handleClose();
-  };
 
   return (
     <>
@@ -96,7 +71,7 @@ export default function ModalRegistro({ open, setOpen }) {
         aria-labelledby="modal-modal-title"
         aria-describedby="modal-modal-description"
       >
-        <Box sx={style1}>
+        <Box sx={styleDsktop}>
           {OpenRegistroPadre === false &&
             OpenLogin === false &&
             OpenRegistroColegio === false && (
@@ -109,8 +84,6 @@ export default function ModalRegistro({ open, setOpen }) {
                     flexDirection: "column",
                   }}
                 >
-                 
-
                   <div
                     style={{
                       display: "flex",
@@ -118,227 +91,67 @@ export default function ModalRegistro({ open, setOpen }) {
                       gap: "2vh",
                     }}
                   >
-                 <div
-                  style={{display:'flex', gap:'2vh', width:'100%',flexDirection:'row', justifyContent:'flex-start', alignItems:'center' }}
-                 >
-                 <IconSch/> 
-                    <Button
-                
-                      transparent
+                    <div
+                      style={{
+                        display: "flex",
+                        gap: "2vh",
+                        width: "100%",
+                        flexDirection: "row",
+                        justifyContent: "flex-start",
+                        alignItems: "center",
+                      }}
                     >
-                      {" "}
-                      <Link
-                        // className={`${style.p} hover-underline-animation`}
-                        style={{ fontWeight: "600",fontSize:'1.4vh', fontFamily: "Poppins" }}
-                        to={"/enroll"}
+                      <IconSch />
+                      <Button transparent>
+                        {" "}
+                        <Link
+                          // className={`${style.p} hover-underline-animation`}
+                          onClick={handleClose}
+                          style={{
+                            fontWeight: "600",
+                            fontSize: "1.4vh",
+                            fontFamily: "Poppins",
+                          }}
+                          to={"/enroll"}
+                        >
+                          Colegio
+                        </Link>
+                      </Button>
+                    </div>
+
+                    <div
+                      style={{
+                        display: "flex",
+                        gap: "2vh",
+                        width: "100%",
+                        flexDirection: "row",
+                        justifyContent: "flex-start",
+                        alignItems: "center",
+                      }}
+                    >
+                      <IconUser />
+                      <Button
+                        onClick={() => setOpenRegistroPadre(true)}
+                        transparent
+                        sx={{
+                          fontWeight: "600",
+                          fontSize: "1.4vh",
+                          fontFamily: "Poppins",
+                        }}
                       >
-                      Colegio
-                      </Link>
-                    </Button>
-                 </div>
-                  
-                 <div
-                  style={{display:'flex', gap:'2vh', width:'100%',flexDirection:'row', justifyContent:'flex-start', alignItems:'center' }}
-                 >
-                   <IconUser/>
-                   <Button
-                      onClick={() => setOpenRegistroPadre(true)}
-                     transparent
-                      sx={{ fontWeight: "600",fontSize:'1.4vh', fontFamily: "Poppins" }}
-                    >
-                     Padre de Familia
-                    </Button>
-                 </div>
-                   
+                        Padre de Familia
+                      </Button>
+                    </div>
                   </div>
                 </div>
               </>
             )}
 
           {OpenRegistroPadre && (
-            <>
-              <div
-                style={{
-                  width: "100%",
-                  display: "flex",
-                  justifyContent: "center",
-                  alignItems: "center",
-                }}
-              >
-                {/* <img style={{ width: "30vh" }} src={Logo} /> */}
-              </div>
-              <form
-                onSubmit={handleSubmit(OnSubmit)}
-                className={style.formLayout}
-              >
-                {/* <div className={style.h1_div}> 
-
-                  <h1 style={{fontSize:'1.8'}}>Completa tussss datos</h1>
-                </div> */}
-                <div
-                  style={{
-                    width: "100%",
-                    padding: "1vh",
-                    display: "flex",
-                    justifyContent: "center",
-                    alignItems: "center",
-                  }}
-                >
-                  <Typography
-                    variant="h5"
-                    sx={{
-                      fontWeight: "700",
-                      fontFamily: "Poppins",
-                      color: "#0061DF",
-                    }}
-                  >
-                    Completa tus datos
-                  </Typography>
-                </div>
-
-                <div className={style.form}>
-                  <div className={style.divInputs}>
-                    <label className={style.label}>Nombre</label>
-                    <input
-                      placeholder="Introduzca su nombre "
-                      {...register("name", {
-                        required: true,
-
-                        maxLength: 100,
-                      })}
-                      className={style.input}
-                    />
-
-                    {errors.name && (
-                      <p className={style.p}>Introduzca su nombre.</p>
-                    )}
-                    <label className={style.label}>Apellido</label>
-                    <input
-                      placeholder="Introduzca su apellido "
-                      {...register("lastname", {
-                        required: true,
-                        maxLength: 100,
-                      })}
-                      className={style.input}
-                    />
-
-                    {errors.lastname?.type === "required" && (
-                      <p className={style.p}>Introduzca su apellido.</p>
-                    )}
-                    {errors.lastname?.type === "maxLength" && (
-                      <p className={style.p}>Demasiados caracteres.</p>
-                    )}
-                    <label className={style.label}>Email</label>
-                    <input
-                      placeholder="Introduzca su correo electronico "
-                      {...register("mail", {
-                        required: true,
-
-                        maxLength: 100,
-                        pattern: /\S+@\S+\.\S+/,
-                      })}
-                      className={style.input}
-                    />
-                    {errors.mail?.type === "required" && (
-                      <p className={style.p}>Introduzca su mail.</p>
-                    )}
-                    {errors.mail?.type === "pattern" && (
-                      <p className={style.p}>El formato es examp@sds.com</p>
-                    )}
-                    {errors.mail?.type === "maxLength" && (
-                      <p className={style.p}>Demasiados caracteres.</p>
-                    )}
-                  </div>
-                  <div className={style.divInputs}>
-                    <label className={style.label}>Telefono</label>
-                    <input
-                      type="number"
-                      placeholder="Introduzca numero de telefono"
-                      {...register("phone", { required: true })}
-                      className={style.input}
-                    />
-                    {errors.phone && (
-                      <p className={style.p}>Introduzca su telefono .</p>
-                    )}
-
-                    <label className={style.label}>Contraseña</label>
-                    <div className={style.DivPass}>
-                      {seePassword === true ? (
-                        <BsEye
-                          onClick={ToggleSeePass}
-                          className={style.Password}
-                        />
-                      ) : (
-                        <BsEyeSlash
-                          onClick={ToggleSeePass}
-                          className={style.Password}
-                        />
-                      )}
-
-                      <input
-                        placeholder="Contraseña"
-                        type={seePassword === true ? "text" : "password"}
-                        {...register("password", {
-                          required: true,
-                          maxLength: 100,
-                        })}
-                        className={style.input}
-                      />
-                    </div>
-                    {errors.password?.type === "required" && (
-                      <p className={style.p}>Campo requerido</p>
-                    )}
-                  </div>
-                </div>
-                <div className={style.divButton}>
-                  <button type="submit">REGISTRARSE</button>
-                </div>
-
-              
-                <div
-                  style={{
-                    display: "flex",
-                    width:'100%',
-                    flexDirection:'column',
-                    alignContent:'center',
-                    justifyContent:'center',
-                 
-                    alignItems: "center",
-                  }}
-                >
-                    <p className="  text-sm">Prefiero iniciar sesion con</p>
-                <div
-                  style={{
-                    display: "flex",
-                    width:'100%',
-                    flexDirection:'row',
-                    alignContent:'center',
-                    justifyContent:'center',
-                    gap: "1.5vh",
-                    padding: "1vh",
-                    alignItems: "center",
-                  }}
-                
-                >
-                    <FB/>
-                  <Gmail />
-                </div>
-                
-                </div>
-
-                <div className={`${style.divButton}`}>
-                  <p>Ya tienes cuenta ? </p>
-                  <div>
-                    <p
-                      onClick={handlerOpenLogin}
-                      style={{ color: "blue", cursor: "pointer" }}
-                    >
-                      Inicia Sesión{" "}
-                    </p>
-                  </div>
-                </div>
-              </form>
-            </>
+            <ModalRegistroFamilia
+              open={OpenRegistroPadre}
+              setOpen={setOpenRegistroPadre}
+            />
           )}
           {OpenRegistroColegio && OpenLogin === false && (
             <FormInscripcion handlerOpenLogin={setOpenLogin} />
@@ -351,6 +164,104 @@ export default function ModalRegistro({ open, setOpen }) {
               )}
           </div>
         </Box>
+
+        {/* <div className={style.divMobile}>
+          <Box sx={styleDsktop}>
+            {OpenRegistroPadre === false &&
+              OpenLogin === false &&
+              OpenRegistroColegio === false && (
+                <>
+                  <div
+                    style={{
+                      display: "flex",
+                      gap: "2vh",
+                      padding: "1vh",
+                      flexDirection: "column",
+                    }}
+                  >
+                    <div
+                      style={{
+                        display: "flex",
+                        flexDirection: "column",
+                        gap: "2vh",
+                      }}
+                    >
+                      <div
+                        style={{
+                          display: "flex",
+                          gap: "2vh",
+                          width: "100%",
+                          flexDirection: "row",
+                          justifyContent: "flex-start",
+                          alignItems: "center",
+                        }}
+                      >
+                        <IconSch />
+                        <Button transparent>
+                          {" "}
+                          <Link
+                            // className={`${style.p} hover-underline-animation`}
+                            style={{
+                              fontWeight: "600",
+                              fontSize: "1.4vh",
+                              fontFamily: "Poppins",
+                            }}
+                            to={"/enroll"}
+                          >
+                            Colegio
+                          </Link>
+                        </Button>
+                      </div>
+
+                      <div
+                        style={{
+                          display: "flex",
+                          gap: "2vh",
+                          width: "100%",
+                          flexDirection: "row",
+                          justifyContent: "flex-start",
+                          alignItems: "center",
+                        }}
+                      >
+                        <IconUser />
+                        <Button
+                          onClick={() => setOpenRegistroPadre(true)}
+                          transparent
+                          sx={{
+                            fontWeight: "600",
+                            fontSize: "1.4vh",
+                            fontFamily: "Poppins",
+                          }}
+                        >
+                          Padre de Familia
+                        </Button>
+                      </div>
+                    </div>
+                  </div>
+                </>
+              )}
+
+            {OpenRegistroPadre && (
+              <ModalRegistroFamilia
+                open={OpenRegistroPadre}
+                setOpen={setOpenRegistroPadre}
+              />
+            )}
+            {OpenRegistroColegio && OpenLogin === false && (
+              <FormInscripcion handlerOpenLogin={setOpenLogin} />
+            )}
+            <div>
+              {OpenLogin === true &&
+                OpenRegistroPadre === false &&
+                isAuth === false && (
+                  <FormLogin
+                    setOpenLogin={setOpenLogin}
+                    OpenLogin={OpenLogin}
+                  />
+                )}
+            </div>
+          </Box>
+        </div> */}
       </Modal>
     </>
   );
